# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "media_engine",
    "description": "Universal media-processing engine — typed artifacts, composable operations, pluggable backends.",
    "version": "0.5.0"
  },
  "paths": {
...
    "/plugins/catalog": {
      "get": {
        "summary": "Get Plugins Catalog",
        "operationId": "get_plugins_catalog_plugins_catalog_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Put Plugins Catalog",
        "description": "Replace the catalog gate. Unknown op / backend keys are ignored\nsilently — the operator might be hiding entries from an extra\nthey're about to install, and writing those keys would surprise\nthem with a 400. The runtime filter just doesn't fire on unknown\nkeys.",
        "operationId": "put_plugins_catalog_plugins_catalog_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/storage/gc": {
      "post": {
        "summary": "Post Storage Gc",
        "description": "Workdir sweep + (optional) LRU eviction.\n\nMirror of ``med storage gc``. When ``apply=False`` the route\nreports what would happen without writing — workdir candidates are\nlisted but not deleted, and eviction runs in ``dry_run=True``\nmode. ``eviction_enabled=False`` in engine config disables the\neviction pass regardless of the ``evict`` flag (the CLI honours\nthe same gate).",
        "operationId": "post_storage_gc_storage_gc_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GCRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GCResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CatalogResponse": {
        "properties": {
          "ops": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ops"
          },
          "backends": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Backends"
          },
          "hidden_ops": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hidden Ops"
          },
          "hidden_backends": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hidden Backends"
          }
        },
        "type": "object",
        "required": [
          "ops",
          "backends",
          "hidden_ops",
          "hidden_backends"
        ],
        "title": "CatalogResponse",
        "description": "Visibility state + the universe of toggleable entries.\n\n``ops`` is every registered op name; ``backends`` is every\n``op.name__backend.name`` key. ``hidden_ops`` / ``hidden_backends``\nare the currently-hidden subsets (from ``plugins.toml``). The Web\nUI renders one checkbox per entry and POSTs back the inverted\nstate via PUT /plugins/catalog."
      },
      "CatalogUpdate": {
        "properties": {
          "hidden_ops": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hidden Ops"
          },
          "hidden_backends": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hidden Backends"
          }
        },
        "type": "object",
        "title": "CatalogUpdate"
      },
...
      "GCRequest": {
        "properties": {
          "apply": {
            "type": "boolean",
            "title": "Apply",
            "default": false
          },
          "sweep_workdirs": {
            "type": "boolean",
            "title": "Sweep Workdirs",
            "default": true
          },
          "evict": {
            "type": "boolean",
            "title": "Evict",
            "default": true
          }
        },
        "type": "object",
        "title": "GCRequest",
        "description": "``apply=False`` is the default — preview what *would* happen."
      },
      "GCResponse": {
        "properties": {
          "applied": {
            "type": "boolean",
            "title": "Applied"
          },
          "workdirs_swept": {
            "type": "integer",
            "title": "Workdirs Swept"
          },
          "workdir_candidates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Workdir Candidates"
          },
          "eviction_enabled": {
            "type": "boolean",
            "title": "Eviction Enabled"
          },
          "bytes_before": {
            "type": "integer",
            "title": "Bytes Before",
            "default": 0
          },
          "bytes_after": {
            "type": "integer",
            "title": "Bytes After",
            "default": 0
          },
          "evicted_artifact_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evicted Artifact Ids"
          }
        },
        "type": "object",
        "required": [
          "applied",
          "workdirs_swept",
          "workdir_candidates",
          "eviction_enabled"
        ],
        "title": "GCResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}